+++ common/src/python/curator/form_curator.py
@@ -3,7 +3,13 @@ import logging
 from flywheel import Client
 from flywheel.models.file_entry import FileEntry
 from flywheel.models.subject import Subject
+<<<<<<< HEAD
 from nacc_attribute_deriver.attribute_deriver import AttributeDeriver
+=======
+from flywheel.rest import ApiException
+from keys.keys import GearTags
+from nacc_attribute_deriver.attribute_deriver import AttributeDeriver, ScopeLiterals
+>>>>>>> main
 from nacc_attribute_deriver.symbol_table import SymbolTable
 from nacc_attribute_deriver.utils.scope import ScopeLiterals
 
@@ -15,9 +21,14 @@ log = logging.getLogger(__name__)
 class FormCurator(Curator):
     """Curator that uses NACC Attribute Deriver."""
 
-    def __init__(self, sdk_client: Client, deriver: AttributeDeriver) -> None:
+    def __init__(self,
+                 sdk_client: Client,
+                 deriver: AttributeDeriver,
+                 force_curate: bool = False) -> None:
         super().__init__(sdk_client)
+        self.__sdk_client = sdk_client
         self.__deriver = deriver
+        self.__force_curate = force_curate
 
     def apply_curation(self, subject: Subject, file_entry: FileEntry,
                        table: SymbolTable) -> None: